requirements.txt
streamlit-pydantic
pydantic
psycopg2-binary
python-dotenv
datetime